refactor UDF
