+++ src/main/resources/application.conf
@@ -12,10 +12,12 @@ default {
       warehouse="warehouse/"
       storage="temperatures/"
       temperature="temperature.parquet"
+      temperature="dashboard.parquet"
       dataPath=${default.hdfs.path.base}${default.hdfs.path.test}
       failPath=${default.hdfs.path.base}${default.hdfs.path.fail}
       warehousePath=${default.hdfs.path.base}${default.hdfs.path.warehouse}
       temperaturePath=${default.hdfs.path.warehousePath}${default.hdfs.path.storage}${default.hdfs.path.temperature}
+      dashboardPath=${default.hdfs.path.warehousePath}${default.hdfs.path.storage}${default.hdfs.path.dashboard}
     }
   }
   xml {
